+++ providers/xai/src/ai/index.ts
@@ -6,8 +6,13 @@
 
 // organize-imports-ignore
 
-export { XAI_ALLOWED_HOSTS, XAI_DEFAULT_BASE_URL } from "./common/Xai_Client";
+export {
+  XAI_ALLOWED_HOSTS,
+  XAI_DEFAULT_BASE_URL,
+  getXaiReasoningEffort,
+} from "./common/Xai_Client";
 export * from "./common/Xai_Constants";
+export { xaiEffortPolicy } from "./common/Xai_EffortPolicy";
 export * from "./common/Xai_ModelSchema";
 export * from "./common/Xai_ModelSearch";
 export * from "./registerXai";