+++ providers/xai/src/ai/runtime.ts
@@ -14,5 +14,6 @@
 // organize-imports-ignore
 
 export * from "./common/Xai_Client";
+export * from "./common/Xai_EffortPolicy";
 export * from "./registerXaiInline";
 export * from "./registerXaiWorker";